Split sidebar loading skeleton helpers
